chores: consolidate vm name validation
